pyproject.toml
[project]
name = "grokking-mnist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autoattack",
    "cupy>=13.5.1",
    "ffcv>=1.0.2",
    "hydra-core>=1.3.2",
    "ipykernel>=6.30.0",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "matplotlib-fontja>=1.1.0",
    "omegaconf>=2.3.0",
    "seaborn>=0.13.2",
    "torch>=2.7.1",
    "torchattacks>=3.5.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "wandb>=0.21.0",
]

[tool.uv.sources]
autoattack = { git = "https://github.com/fra31/auto-attack" }
